feat: add redux persist
