+++ MyApp/src/store/index.ts
@@ -1,11 +1,37 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { setupListeners } from '@reduxjs/toolkit/dist/query';
 import { tasksApi } from './features/api/task';
+import { combineReducers } from 'redux';
+import {
+  persistReducer,
+  FLUSH,
+  REHYDRATE,
+  PAUSE,
+  PERSIST,
+  PURGE,
+  REGISTER,
+} from 'redux-persist';
+import storage from '@react-native-async-storage/async-storage';
+
+const reducers = combineReducers({
+  [tasksApi.reducerPath]: tasksApi.reducer,
+});
+
+const persistConfig = {
+  key: 'root',
+  storage,
+};
+
+const persistedReducer = persistReducer(persistConfig, reducers);
 
 export const store = configureStore({
-  reducer: { [tasksApi.reducerPath]: tasksApi.reducer },
+  reducer: persistedReducer,
   middleware: getDefaultMiddleware =>
-    getDefaultMiddleware().concat(tasksApi.middleware),
+    getDefaultMiddleware({
+      serializableCheck: {
+        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
+      },
+    }).concat(tasksApi.middleware),
 });
 
 setupListeners(store.dispatch);
